Added items into sidebar menu
Rathern than try to repurpose the homepage menu, the data for the items
was extracted and placed into a new data directory, and then the options
were created separately. The sidebar options were placed into the layout
folder, since some styling transitions ocurr when the sidebar is hovered
over. Part of the difficulty comes form the options container having no
width, so it has had to be transformed by a fixed amout which is not
ideal but works.