Complete restyle
Mainly focused on home page. Added new font which is casul and playful
rather than sleek and serios fonts used before. I think this is more the
vibe I want to go for, and is more me. Simplified main page content with
focus on site name an an underline that will in future be clickable and
will change. An orange dot has been added in main title, which is a
design decision that makes the title stand out and adds a themeing
choice that can be consistent through the rest of the site. ORange is
also a fun and playful colour, which again is what I want to reflect
with the site since it is personal to me.

Also disabled layout since will retheme toolbar later, and also do not
want it on main page.