Completely overhauled structure of site
Decided to do a single scrollable apge with anchor sections instead of
multiple pages. THis resolves the issue of directly navigating to a page
via url with react router. It also is better for SEO since all the
content will be available.

I think it also looks quite nice and simplifies the site.